Fix a bunch of indentation-related issues
Previously, if a to-do was unindented while at indentation level zero,
it would move to the end of a list. This was never meant to happen and
has been restricted to indentation levels greater than zero.

To indicate better that an attempted change of the indentation level
has been rejected, a short tween is played in those cases now.

Also, while the old code correctly specified `allow_echo` as `false` for
the "indent_todo" and "unindent_todo" actions, that unfortunately meant
the echo events themselves were ignored (and thus handled as a regular
attempt to focus the next/previous UI element). To work around this, two
new cases were added, which only trigger if the regular ones didn't.