Apply current indentation to MainRow as well
That is, the panel of an indented heading will appear visually shortened
at the start now, relative to the current indentation level.

The panel with the EditingOptions won't be shortened, to ensure all the
options fit in. However, its Triangle will align with the CheckBox (or
FoldHeading) of the currently edited item now again. (This *did* work
previously, but was broken since the scaling changes in b0bb7c9.)

Increases the indentation width slightly (from 20 to 25 pixels).

Also, makes sure the ExtraInfo and FoldHeading are hidden by default, to
avoid text misalingments between headings and regular to-dos.