Add `show_outside_month_dates` setting
If enabled, and the current month doesn't start with the first, or end
with the last day of the week, the resulting gaps in the calendar widget
will be filled up with dates from the previous or next month.

Disabled by default. Changes persist across restarts of the app.

Also, fixes a bug where the explanation text of the "FirstWeekDay"
setting was no longer justified, after toggling the setting once.