Add collapsible categories to the Settings panel
Each category marks an instance of the new `collapsible_category` scene,
which is supposed to be used with "Editable Children" enabled. Then, the
categories name can be set by adding the text of its "Heading" node, and
the visibility setting of its "Content" node determines the current state
(collapsed or not), which can later be toggled by clicking the Heading.

Right now, there are 4 different categories:
- "General" (uncollapsed by default)
- "List View" (collapsed by default)
- "To-Dos" (collapsed by default)
- "Calendar" (collapsed by default)

Also, modifies the Explanations of some settings slightly.