Fix hidden mouse cursor not re-appearing again
When c1edf1c introduced auto-hiding the mouse cursor while typing in a
to-do's text, there was one oversight: The cursor was made visible again
from the to-do's *own* `_input` function. This meant, that when the item
was the only visible one in the view, and got deleted without using the
mouse (i.e. using the keyboard shortcut, or by submitting an empty text)
the mouse cursor would remain hidden even when moving the mouse.

This commit fixes that by moving the relevant code for making the cursor
visible to `list_view.gd` instead. That also avoids unnecessary function
calls (previously, *every* item would react to an input event).