Simplify the handling of the LineHighlight
... by directly including upper- and lower-padding in each to-do item.
That is, each item is now *exactly* as high as one row in the DayView,
there's no longer any need for the TodoList to have a separation value
and a vertical padding around itself to ensure a proper result.

When adding a new to-do by left clicking, that's now directly using the
position of the LineHighlight, instead of the event's position.