Refactor EditingOptions into separate scene
This significantly reduces the number of nodes in the scene tree at any
given time, as previously all to-dos held their own version in memory,
whereas now only the currently needed one will be part of the tree.

Fixes a bug where calling an item's `edit` function would actually call
it twice, as the function grabbed focus and the `focus_entered` signal
was connected to the `edit` function as well.

Also, fixes a bug in the calculation of the `_shrink_threshold`, which
previously only considered the longer version of the Bookmark label, but
not the longer form of the Delete label (if a to-do is a heading).