Accept input action echos in the editing options
Without this, holding down any of the shortcuts, e.g. CTRL+B for making
the text bold, made the item appear bold (for the first event) but then
would add a bunch of B's to the to-do's text (for every echo event).

Since Godot's shortcut system doesn't provide a way to deal with echos,
this commit replaces them with a manual check in the editing option's
`_input` function: It first checks with `allow_echo` set to false, which
only triggers for the initial press, and then *again*, with `allow_echo`
set to true, which merely accepts the event, but does nothing.

As the button's no longer have shortcuts now, the input action instead
needs to be manually set via an export variable in each tooltip.