Rework how side panels are spawned and hidden
Previously, all possible variants of the side panel were part of the
scene, and only made visible/invisible appropriately. While this worked,
it also meant each variant was still processed in the background, which
e.g. would result in an error, when searching for a to-do's text while
the "Capture" panel wasn't open (but still in the scene tree!).

The new solution prevents that by adding and removing the respective
variants from the tree, instead of only togggling their visbility.

Also, establishes a `_set_initial_state` function in `bookmarks.gd`.