Reduce reliance on using `get_parent()`
This commit should make the project much less prone to breaking after
a change to the tree hierarchy (i.e. by adding/removing a node).
That is achieved by introducing some helper functions that move up the
tree until the requested node is found or the root node is reached.

Where this was not sensible, a helper variable was introduced that holds
the value returned by `get_parent()`. While this might still break, at
least it will only require one change to fix the issue that way.

In other places the calls could be avoided entirely.