Fix moving empty to-do items via drag & drop
Previously, moving an empty (e.g. newly created) to-do item resulted in
an error. This is fixed now, since moving a to-do won't auto-submit its
text automatically, or release its focus any longer.

This is complemented by switching the DragHandle's focus_mode to "None"
(so clicking it doesn't steal focus either), and reentering a to-do's
edit mode in `drop_data`, if it was active before the move.

In the drag preview, the EditingOptions will be hidden.

Also, the `_strip_text` portion of `_on_edit_text_submitted` has been
moved into its own function now, which is also called when a to-do is
saved to disk. That ensures it's no longer possible to force quit the
app to create a to-do text with unwanted elements (like spacing).