Change tooltip background color
... from neutral_4 to neutral_5 to ensure it always contrasts its
surrounding (e.g. the panel's background color for a heading).

Since the tooltip node is not a Control node, it stops the automatic
inheritance of the theme, and this will only work if the theme of the
tooltip_panel is applied manually (which it is now!).

Also, adds tooltips for the FoldHeading and CheckBox toggles of a to-do.
Their popup delay is deliberately higher than for other tooltips so they
don't obstruct the most common interactions with a to-dos.