Fix restoring a folded heading on startup
Previously, only the `is_folded` state of the heading was restored, but
all of its subordinate items would become visible again.

This commits fixes that by adjusting `_apply_state_relative_formatting`
to only make a to-do visible again when it's *not* folded, setting the
`is_folded` property to true on folded sub items, and returning early
from the `is_folded` setter function if it's not heading.

As a side effect, this also drastically cuts down on the number of times
the `folded` and `unfolded` signals are emitted, as previously they were
emitted for *every* sub item as well (which they now aren't anymore).

Also, makes sure that a folded to-do (and all its sub items) are properly
unfolded again when its `is_heading` property is reset to false.

(The changes to `export_presets.cfg` all stem from the update to Godot
4.4 - I hadn't exported a current build until that point!)