Let calendar month & year react to hover events
That is, it's no longer necessary to click the respective buttons to
change the currently shown month or year in the calendar widget.

This required manually making no item in the popup menu focused when
clicking the button - in Godot 4.3 that was still the default.

Also, fixes an alignment issues with the YearSpinBox, introduced by a
recent theme change, that made the widget grow slightly on hover.