Fix several issues with the calendar widget
There is a short delay of 0.4 seconds now after hovering with the mouse
over the month name or year before they become editable.

Clicking on the month button will show the popup menu right away, while
clicking the button another time will hide it again.
Previously, clicking it while the popup menu was open would close the
calendar widget - with "embedded subwindows" enabled in the project
settings that seems to not happen anymore, though.

The brief flicker of the popup when moving the mouse from the button to
the menu and back is now gone. Moving the cursor away from the button
or popup menu will still close the popup menu (if opened).