Add "SubItems" container to to-do items
That is, sub items are now *actually* children of their parent to-do,
not just visually distinct items in the same container. This greatly
simplifies the logic required to implement certain features.

This is a WIP, and many of the previous features (folding, moving, etc.)
are still broken, which is why this is refactored on a new branch.