Fix adding new to-dos inside a SubItems container
By connecting to the item list's `gui_input` signal, accepting any input
events of the left mouse button, and calling a local `add_todo` function
in `item_list.gd`. Note, that this function is not yet on par with the
original function in todo_list.gd (which still exists).

Removes the LowerPadding in a to-do item and adds it to the UpperPadding
instead. This avoids different index results between clicking slightly
above or below a line in a to-do list.