Reimplement most todo item keyboard actions
That is: Adding a todo above or below, moving a todo up or down one
entry in its item list (but not outside of it!), and selecting the
todo's direct predecessor or successor (if there is one).

With the old implementation, I tried very hard to follow the "call down,
signal up" advice. However, that requires reconnecting all those signals
every time a to-do is moved to a different list, which is why I gave up
on that with the rewrite, and simply call up using `get_item_list` now.
Since a todo-item is only supposed to be used in this one context (with
a item list above it) this approach seems sufficiently robust.

Also, adapts the item list's `add_todo` function to work with negative
indices (to add relative to the end of the list), as well. The new
parameter `auto_edit` can be set to false to disable entering the
to-do's edit mode right away after creation.

Those changes, along with the "Items" container's vertical size flags
changed to "Fill", and `show_line_highlight` being capped at the
combined minimum size of the Items list, allowed for getting rid of the
`add_todo` function in `todo_list.gd` completely.