Add sub item folding
That is, while a to-do has (at least) one sub item, it's CheckBox is
replaced by the FoldHeading toggle. Clicking it hides all sub items.

The ExtraInfo panel was moved to the right side of the item, and only
contains the number of *direct* sub items and no indication of whether
those are already marked as done or not.