Adjust getting successor-/predecessor-to-dos
It is now possible to get items outside of the currently selected item
list, that is, reach every item in the list via the arrow keys.

Introduces a new helper function `get_sub_item` to `todo_item.gd`, which
acts as a shortcut for `get_node("%SubItems").get_child()`.

Also, stores the value of `caret_column` when submitting a to-do's text,
and restores it again when `edit` is called via a meta variable.