Fix some non-working save requests
First, for indenting or unindenting a to-do, which still relied on the
(no longer existing) `list_save_requested` signal of the item.

Secondly, for deleting a to-do, which errorneously tried to get the
item's to-do list by calling `get_to_do_list` *after* the item exited
the tree, which made the function fail.