Stop unfolding subitems when reparenting a to-do
When reparenting a to-do (either because it was moved to another to-do
list, or because it was (un)indented to another item list), that would
cause `child_entered_tree` being emitted for each sub item again.
Since `_on_sub_item_added` (which is connected to `child_entered_tree`)
sets `is_folded` to false, that meant the sub items of a folded item got
unfolded again automatically when moving or (un)indenting it.

However, that behavior was only ever meant to be relevant when indenting
a to-do (so it wouldn't become invisible by being moved into a tem list
that was currently folded away)! This commit removes the respective line
from `_on_sub_item_added` and directly adds it to `indent_todo` instead.