Move away from directly setting indentation_level
The variable still exists and is used by the item list functions
`indent_todo` and `unindent_todo` to ensure the maximum indentation
level (of 3) isn't exceeded. All other mentions of the variable are
either removed, or reworked to route through `item_list.gd` now.

When adding a new to-do below another one with sub items, the new to-do
will be assumed to be a sub item of that other to-do by default. When
adding multiple items in a row, this should require fewer manual changes
in indentation by the user then the alternative of assuming every newly
added item to be an unindented one at the same level.

Adds a new helper method `has_sub_items` to `todo_item.gd`.