Uncheck an item when removing its last sub item
This is highly opionated, but it felt wrong to me that an item that only
had its state set to States.DONE due to the state of its sub item, would
retain that state once said sub item got removed or unindented.
If the user intended that, they need to manually toggle the item's state
back now, but I'd argue that's still better than overlooking a remaining
to-do because its state was erroneously set to States.DONE.

Also, applies the modulation in `_apply_state_relative_formatting` to
Toggle (i.e. FoldHeading AND CheckBox) instead of just CheckBox.