Allow dropping a dragged to-do on another to-do
That will add the respective to-do to the other to-do's `SubItems` list,
unless it's already part of that, in which case it will be moved to the
end of said list instead (if it isn't already there).

This undos the changes made in 80e7596, while still allowing the user to
turn one to-do into another to-do's sub item via drag & drop only.

Also, moves the early return checks from the `_drop_data` functions to
`_can_drop_data` instead, which has the added benefit of visual feedback
(the cursor changes to FORBIDDEN when the method returns false).