Set the "Indentation" mouse filter to "Stop"
This also stops mouse inputs from propagating further up in the tree,
which breaks the old way of showing and hiding the line highlight via
`mouse_entered` and `mouse_exited` in `scrollable_todo_list.gd`. The
relevant portions are removed and the signals are instead connected to
those functions from within `item_list.gd` to restore the ability.

Changes the LineHighlight's alpha modulation to 1.0 or 0.0 instead of
calling `show()` and `hide()`, as the latter will reset the highlight
position to the first line due to it being the child of a container.
This also makes it obsolete to wait one frame before calculating the
line's vertical position in `show_line_highlight`.