Fix a bug when hovering an empty to-do list
Since the last commit connected the `mouse_entered|exited` signals to
callbacks that relied on the first item to access its to-do list (and
through that it's `show_line_highlight` method), those callbacks failed
when the entire to-do list was empty and had no first item.
Instead they  now use a local copy of the `get_to_do_list` method.

Also, sets the default cursor shape of the "SubItems" container to
"Pointing Hand" to match the one set for the to-do list.