Temporarily unfold sub items matching a search
Previously, when a folded item contained a sub item, that matched the
current search term, there was no visual indication for that: the sub
item changed its color, but since it was hidden it didn't matter.

Now, when a sub item matches the current query, and its parent todo is
folded (!), it will turn its parent's sub item list visible on its own,
which is undone when the search query no longer matches or gets cleared.