Fix nested sub items not unfolding on search match
The code in c7a3664 did not account for situations like this:

- Item A (unfolded)
  - Item B (folded)
    - Item C (unfolded)
      - Item D (unfolded)

When searching for "D", the last item did *not* get shown again, since
the `is_folded` state of its direct parent C was still `false`.

This commit adresses that by iterating through *all* parent to-dos of a
to-do that matches the current query and is not visible in the tree. It
also adapts the code in the setter function of `is_folded` accordingly,
and refactors it into its own method `contains_search_query_match`.