Fix "Jump To" button for bookmarked sub items
... by connecting the `bookmark_jump_requested` signal to the to-do list
instead of each to-do item individually, and then traversing its items
from the top via the new `get_item_for_line_number` function.

The old way of doing this failed because it assumed Godot's `get_index`
function matched the line number of the to-do in question – which was no
longer the case with the new SubItems container, though.