Fix a bug in `get_line_number_for_item`
Starting with a5d38ea, the method no longer considered properly that the
method might reach the end of the item list *without* finding the item.
There it would (correctly) return its current value of `i`, which then
was (incorrectly) treated as the final result of the function.

This commit fixes this by returning `i` with a negative sign in that
case (which can't happen otherwise) and then checking if the result is
below zero or not when returing from a recursive function call.