Update sub item bookmarks when moving their parent
Previously, the bookmarks were only updated when moving the bookmarked
items themselves. If they were sub items of another item and said item
was moved, that meant the bookmark now pointed to a wrong to-do.

This commit fixes that by adding a `update_bookmarked_items` function to
`item_list.gd`, which (recursively) runs through all the sub items of a
to-do (and their sub items, and so on) and emits `bookmark_changed` when
any of those sub items is a bookmark (i.e. `is_bookmarked` is true).