Focus JumpTo button on BookmarkIndicator click
Clicking the bookmark indicator will no longer edit the to-do, but
instead focus the "JumpTo" button of the respective bookmark in the
bookmarks panel (which, since last commit, is automatically brought
after clicking the bookmark indiator).
This makes it easier for the user to spot the connection, while still
allowing an edit by either pressing Enter (since the JumpTo button is
focused now) or clicking on the button (which always works).