Fix `_on_items_child_order_changed`
The previous version of this method only considered direct children of
`Items`, but not the possibility of sub items (i.e. nested item lists).
It also still used `get_index` instead of the new `get_list_index`.

Fixing this means it's no longer necessary to check if a moved item or
any of its sub items are bookmarked and update their bookmarks, as that
will be done when running `_on_items_child_order_changed` anyway. What
the previosu solution did not consider was the fact that a bookmarked
item will *also* require an update if one moves another item before it
(as that will push the bookmarked item further down the list).