Finetune behavior when dragging a to-do
Firstly, while dragging a to-do over a line between itself and one of
its sub items, the line highlight will no longer be shown. This is only
cosmetic: It wasn't possible to drop items there anyways.

Secondly, while dragging a to-do around, the CheckBox or FoldHeading of
the currently hovered item will no longer change to its "Focused" theme
type variation, as that doesn't provide any useful information. The line
highlight should suffice in those cases, and if a to-do is dragged ontop
of another item, that item's MainRow style will change as well.