Only show ExtraInfo on folded to-dos
Also, include the amount of done items in the ExtraInfo again. That is,
there will now be two numbers "(X/Y)", where X is the amount of to-dos
that are marked as either canceled or done, while Y is the total count
of all to-dos (and their sub items, and so on...).
When "Fade Ticked Off To-Dos" is active, and all sub items of an item
are marked as canceled or done, the ExtraInfo will get modulate to half
of its alpha value (like the Toggle and Content).

Lastly, place the BookmarkIndicator on the left of the DragHandle.