Fix dragging to-dos out of view
That is, dragging a to-do to the side of the window, scrolling the view
until the day panel it's originating from is no longer visible.
Before this commit, a faulty type check in `drag_overlay.gd` prevented
the DragOverlay from appearing. Also, the `is_dragged` variable of the
day panel (which is used to keep the to-do around until the drag ends)
referred to a similarly named variable in `scrollable_todo_list.gd` in
its getter variable, but never updated that variable in a setter. This
commit solves this by simply removing the 2nd variable altogether.

Also, renames the `get_to_do_count` method to `get_done_items_count`,
which was actually meant to be amended to the previous commit...