Avoid superfluous emits of `bookmark_changed`
Previously, submitting a bookmarked item's text, even if it wasn't
changed at all, would call `_on_edit_text_changed` which needlessly
prompted an update of the associated bookmark.

Furthermore, when loading a DayPanel's to-do list for the first time,
the signal was emitted for each item that came *after* the bookmarked
item, even though the bookmarked item's index didn't change (as items
are added from top to bottom when restoring them from disk).