Prevent race conditions when updating bookmarks
Given the following scenario:

Day 1:
- Bookmark A

Day 2:
- Bookmark B

Moving one of those bookmarks above the other led to a race condition
where the bookmark of the moved to-do got updated to the date and list
index of the other bookmark first, which meant the second update, that
was actually meant for the second bookmark, updated the first again.

Similarly, when moving a to-do, `_on_items_child_order_changed` gets
called twice: once for the reparent-call (which puts the to-do at the
end of its item list), and again for the `move_child`-call (which puts
it at its actually intended, final position).

This commit introduces two variables to prevent those cases:
- `_has_requested_bookmark_update` in todo_item.gd
- and `updated_this_frame` in bookmark.gd`.