Fix dropping a bookmark on the capture panel
Previously, dragging & dropping a bookmarked to-do to the capture panel
resulted in a number of errors, as the panel does not have an associated
date, which `_on_bookmark_changed` assumes to exist, though.

This commit fixes that by removing the to-do's bookmark before dropping.
Ideally one would also be able to bookmark to-dos in the capture panel,
but until that's realized this at least prevents any errors.