Avoid superfluous emits of bookmark_changed, pt. 2
The `bookmark_changed` signal is no longer emitted when the `text`
variable is set: For the initial setup it's not necessary (as the
bookmark data is read from disk and already correct), and in all
following cases `_on_edit_text_changed` ran before it.

In `_on_edit_text_changed`, the signal is now emitted in *all* cases
(including the special case where the new text starts with "- ", as
that might be copy pasted and still contain changes after that) and
uses `get_list_index` instead of the regular `get_index` now, too.

In `_on_sub_item_removed`, the signal only gets emitted while its day
panel is *not* queued for deletion. Previously, it also got emitted when
a sub item got removed because it scrolled out of the list view.