Fix bookmark's not being updated correctly
Again.....

First, adds back the "not" in the if-condition of `_on_sub_item_removed`
(introduced by 02b6bf7, then accidentally removed again in e5d513a).

Secondly, adds a `last_date` variable to `todo_item.gd`, which serves
the same purpose as `last_index` but for the date, as that's required
when moving a bookmarked top-level item (i.e. one that's not a sub item
of another item) to another day via drag and drop.