Make DragHandle always visible
... but reduced to 10% of its alpha modulation, until the to-do item is
hovered with the mouse. This means that hovering over an item no longer
moves the BookmarkIndicator and/or ExtraInfo to the left.

Also, the DragHandle now acts as the new indicator of whether an item is
currently hovered with the mouse or not. The CheckBox and/or FoldHeading
are only colored in blue anymore if they are hovered themselves - in all
other cases the DragHandle will be colored blue now instead. The former
takes precedence over the later! (via the Toggle's `mouse_filter`)