Fix focus cancelation not working properly
... when the search bar was focused. (Also, keep track of whether the
mouse cursor is still hovering over the search bar, and if so restore
"SearchBar_Hover" instead of "SearchBar" once the focus is released.)

... when a bookmark's "JumpTo" button was focused (after clicking the
bookmark button in the associated to-do item).

The former has to be explicitly handled via the search bar's `gui_input`
signal, while the latter can be handled in the `unhandled_input` method
of the main window, akin to left clicking anywhere else already did.