Reset MainRow type variation when to-do is moved
Previously, when a top-level to-do (i.e. a to-do at indentation level 0)
with (at least) one sub item was indented, or became sub item of another
to-do, it's theme type variation ("ToDoItem_Heading") remained active –
although that's only supposed to happen for top-level to-dos!

Also, fixes the `indentation_level` variable of to-dos not being updated
when moving them to a different indentation via drag & drop.