Maintain backwards compatibility
... by updating to-do files that still use the old syntax:

1) If a line starts with two spaces, turn them into four.
2) If a line starts with "v " (i.e. it's an unfolded heading), replace
   that with "[ ] " (i.e. a regular to-do) and indent *all* following
   to-dos until the next heading (unfolded or folded) is reached.
3) If a line starts with "> " (i.e. it's a folded heading), replace that
   with "[ ] >" (i.e. a folded to-do) and indent *all* following to-dos
   until the next heading (unfolded or folded) is reached.

While the patch will parse *all* files (including files that already are
compatible with the new syntax), it should *only* rewrite legacy ones.

The (new) setting `sub_items_refactor_patch_applied` ensures that this
patch is only applied once, as the variable defaults to false, and is
set to true after the patch has been applied to all files.

The ` _is_valid_filename` method was changed to also return true if the
filename is equal to "capture.txt", and a specialzed case was added to
`_load_file` to commit the capture to the cached data as well.

With this, the sub-item-refactor branch should be ready to merge!