Fix several scrolling issues
First, fix a bug introduced by e17df12, which not only consumed right
and middle mouse button inputs, but scroll wheel inputs as well! That
made it impossible to scroll a to-do list while hovering the mouse over
a to-do's "Edit" field (which takes up most of a the space).

Secondly, add a 13-pixel high spacer element to the end of each item
list, which forces the scroll container to scroll down until the next
line element, and not just to the end of the to-do's text.

Thirdly, remove the ScrollButton's (line) texture while it's in UP-mode.
Since each TodoList adds a 7-pixel offset to their lines, the next line
was still visible slightly underneath the button, which made it appear
as a duplicate. The offset ensures that the line remains clickable, so
one can add a to-do above the topmost visible one without scrolling.

Lastly, adjust the `_scroll_one_item_up|down` methods to *only* scroll
in full increments of `TODO_ITEM_HEIGHT`. This stops a to-do list from
slightly shifting vertically when scrolling all to the end.