Fix two small regressions
First, the if-condition added to `_on_item_added` with the last commit
was wrong and used "!=" instead of "is not", resulting in headings not
being auto checked off if all there sub items were checked off.

Secondly, in blocking middle mouse button inputs for focusing the Edit
field of a to-do item also rendered using it for pasting the clipboard
content when the item was *already* focused impossible.

This commit fixes both of these issues.