Streamline procedure for adding to-dos
A to-do with sub items is only rendered as a heading now, if its text
ends with a colon (":") character. Consequently, if the user submits a
to-do that ends in a colon, the next to-do will be auto-indented, a new
feature provided by the `add_sub_item` method in `item_list.gd`.

The keybindings for adding a to-do above or below were changed: SHIFT
and ENTER will now add a to-do below, while SHIFT+CTRL+ENTER will add
one above. Unlike before, both of these are input actions now and have
to be exact matches (i.e. no other keys held) to trigger anything.
If the submitted item is a newly created one, only pressing ENTER will
*also* create a to-do below. When editing an item that already has been
submitted before and got its text set, that's not the case though!

Lastly, expands the `_strip_text` function in `todo_item.gd`, to also
account for color tags and markup (like asterisks and - newly added -
underscores). In not only stripping those tags from the text, but also
applying their effects right away, this function can now be reused in
`_load_from_disk` as well, significantly shortening that function.
The `_load_from_disk` now also considers that the loaded text *might*
end up empty after stripping away all those tags, in which case the
resulting to-do will be deleted. This shouldn't happen for files written
by Habituary itself, but may occur nonetheless if the user modifies a
to-do list file with an external program.