Fix a bug with toggling to-dos via keyboard
... where the CheckBox' theme type variation was not updated properly
while `_contains_mouse_cursor` was true. Instead, the code now checks
explicitly if the mouse is over the *CheckBox* and renames the variable
accordingly, since it was only relevant in that context anyway.

Also, removes the (now superfluous) Content node around the Edit node,
and purges the associated type variation from the generated themes.