Align the lines in Capture and ListView
... by having the Capture mirror the bottom margin of the MainPanel
(minus the SidePanels own bottom margin), calculating the remaining
space for the to-do list and then applying the result of that space
modulo the height of a to-do item as the top margin.
If the root window changes size, this will get recalculated!

The calculation of the remaining space contains a final offset of 7
pixels that was required for a correctly looking result, although I
couldn't trace its origin and why it's necessary.