Flatten cache data structure
Instead of storing the (at this point still unused) cache data as a set
of nested dictionaries with year, month and day as keys:

  { "2025": { "06": { "01": { "content": ..., "last_modified": ...}}}}

the script now simply uses the whole date as *one* key:

  { "2025-06-01": { "content": ..., "last_modified": ...}}}}

This eliminates the need for a special case just to handle the capture
file, at the mere cost of not being able to (easily) filter the cached
data down to a specific month or year (which isn't yet required).

This commit also splits the cached content at newlines, ignoring empty
lines, which should make it easier to parse it line-by-line later.