Fix not being able to repeatedly add a to-do above
The previous code did not account for the possibility that one might
press "add_todo_above" on new items as well. So when creating a to-do
above another one, then adding some text to it and trying to add one
more to-do above that one, that to-do was added below regardless!