Fix scrolling the list view during a search
Turns out, ce71cf4 was wrong when stating that `_check_for_search_match`
was "needlessly called". While the search query indeed does not persist
across restarts, there might be an active search query when initializing
a to-do later on: when scrolling the list view during a search.

So this commit undos that change, while trying to make it more obvious
why and when the call to `_check_for_search_match` is necessary.