Save to-dos to and load to-dos from data cache
Previously, a to-do list's data was read directly from disk, every time
the list came into view. That also meant, it was not possible to move a
to-do to a list outside the current view (e.g. next week), as (from the
viewpoint of Habituary) that list did not exist.

Now *all* data is read once the client is started and kept in memory as
a "data" dictionary in the Cache autoload. This enables modifying to-dos
at arbitrary dates, which is required for several planned features.