Update to-do lists modified outside on focus enter
Most of the functionality for this was already in place: The "Cache"
autoload calls `_load_directory` on `NOTIFICATION_WM_WINDOW_FOCUS_IN`,
which will update the cache for any files with changes that are newer
than the cached last modification date.

This commit adds a "content_updated"-signal to the "Cache" autoload,
which is emitted in those cases alongside the updated key. Then all
that was left to do was to connect all to-do lists to that signal, and
(if the key matches the list's cache key) reload the list content.

This solution is not perfect! For example it's possible to modify a file
outside, but delay it's saving somehow, then switching back to Habituary
right away, so the window will get focused *before* the file is saved.
However, it should cover the most common case (the user switching from
the app to another and back) just fine, which seems good enough.