Fix several issues with restoring to-do lists
First, and most importantly, it could happen that `load_from_cache` was
called before all old items got free'd (using `queue_free`). This is now
solved by waiting one frame until the child count equals zero.

Secondly, if a raw to-do list file contained any amount of spaces that
was not a multiple of 4 in front of a to-do, the to-do's text contained
the "[ ] " markup text as well. This is now solved by stripping the left
side of the to-do after indenting it for each set of 4 spaces.

Lastly, when the raw to-do list file contained more sets of four spaces
than what Habituary allows for, one could see the to-do shaking briefly
due to the call `_reject_indentation_change` in `indent_todo`. This is
now solved via an optional `visual_feedback` parameter that defaults to
`true`, but can be set to `false` where this behavior isn't wanted.