Add proof of concept for a global to-do search
This is a work in progress, and not yet feature complete! When searching
for a query via the search bar, one can now press SHIFT+ENTER to switch
the (new) main panel state from LIST_VIEW to GLOBAL_SEARCH.

The search is performed on *all* cached contents, with later dates shown
first (i.e. future before present before past dates). Later it should be
possible to jump to any matching to-do, similar to how it works with any
bookmarked to-dos already. To exit the global search panel and go back
to the list view panel again, one must clear the search bar completely.