Fix "clearing" search query when hitting Enter
22c2fc4 aded an additional if-condition to the 'gui_input` callback to
stop the code from superfluously executing the respective code on both
key press *and* key release, and consequently limited the execution to
the event being released only. However, it *needs* to be key press, as
this is when TextEdit (via the built-in behavior) would enter a newline
when the user hits the Enter key - and we have to prevent this here, as
the search query is supposed to be limited to one line.