Let scheduling an item copy, not move the original
The original is kept in the capture list, and automatically postponed
until one day after the date that it was scheduled for. If it has been
completed then, the user might delete it. If not, they can reschedule it
another time without any need to move the unfinished to-do back.