Add calendar widgets to the review view
... to easily pick *any* date to postpone to or schedule for. This
replaces the previous buttons of "Today", "Tomorrow", "Next Week".

This required a few changes to the calendar widget:
- `_on_visibility_changed` is called in `_set_initial_state` now (to
  update the widget to the correct month).
- Clicking on a `DayButton` no longer changes `Settings.current_day`
  automatically. Instead, the `associated_day` is emitted alongside a
  newly created `day_button_pressed` signal, which replaces the signal
  `closed` used by the overlay in the old implementation.
- `_on_today_pressed` was renamed into `reset_view_to_today` and is now
  considered a public method of the calendar widget.

Picking a date will call `_postpone|schedule_current_item`, then reset
the calendar view to today. Picking dates in the past is *technically*
still possible, but won't trigger `_postpone_schedule_current_item`.

Also, starting with this commit, if the review date of a to-do is equal
to today's date, that review will be considered immediately due!